+++ Neo/Neo/Services/MatrixHighLevel.cs
@@ -19,7 +19,7 @@ namespace Neo.Services
         /// </summary>
         /// <param name="matrix"></param>
         /// <returns></returns>
-        public static double GetDeterminant(Matrix<double> matrix) 
+        public static decimal GetDeterminant(Matrix<decimal> matrix) 
             => matrix.Determinant();
 
         /// <summary>
@@ -27,7 +27,7 @@ namespace Neo.Services
         /// </summary>
         /// <param name="matrix"></param>
         /// <returns></returns>
-        public static double GetRank(Matrix<double> matrix)
+        public static decimal GetRank(Matrix<decimal> matrix)
             => matrix.Rank();
 
         /// <summary>
@@ -35,7 +35,7 @@ namespace Neo.Services
         /// </summary>
         /// <param name="matrix"></param>
         /// <returns></returns>
-        public static Matrix<double> GetReverseMatrix(Matrix<double> matrix)
+        public static Matrix<decimal> GetReverseMatrix(Matrix<decimal> matrix)
             => matrix.Inverse();
         
         /// <summary>
@@ -44,7 +44,7 @@ namespace Neo.Services
         /// <param name="matrix"></param>
         /// <param name="n">Exponentiation value</param>
         /// <returns></returns>
-        public static Matrix<double> Exponentiation(Matrix<double> matrix, int n)
+        public static Matrix<decimal> Exponentiation(Matrix<decimal> matrix, int n)
         {
             var tempMatrix = matrix;
             for (int i = 1; i < n; i++)
@@ -58,7 +58,7 @@ namespace Neo.Services
         /// </summary>
         /// <param name="matrix"></param>
         /// <returns></returns>
-        public static Matrix<double> Transpose(Matrix<double> matrix)
+        public static Matrix<decimal> Transpose(Matrix<decimal> matrix)
             => matrix.Transpose();
     }
 }
